Helper function for test socket creation
