+++ test/Microsoft.AspNetCore.Server.KestrelTests/TestConnection.cs
@@ -29,24 +29,7 @@ namespace Microsoft.AspNetCore.Server.KestrelTests
 
         public void Create(int port)
         {
-            _socket = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp);
-            if (PlatformApis.IsWindows)
-            {
-                const int SIO_LOOPBACK_FAST_PATH = (-1744830448);
-                var optionInValue = BitConverter.GetBytes(1);
-                try
-                {
-                    _socket.IOControl(SIO_LOOPBACK_FAST_PATH, optionInValue, null);
-                }
-                catch
-                {
-                    // If the operating system version on this machine did
-                    // not support SIO_LOOPBACK_FAST_PATH (i.e. version
-                    // prior to Windows 8 / Windows Server 2012), handle the exception
-                }
-            }
-            _socket.NoDelay = true;
-            _socket.Connect(new IPEndPoint(IPAddress.Loopback, port));
+            _socket = CreateConnectedLoopbackSocket(port);
 
             _stream = new NetworkStream(_socket, false);
             _reader = new StreamReader(_stream, Encoding.ASCII);
@@ -141,6 +124,28 @@ namespace Microsoft.AspNetCore.Server.KestrelTests
             var count = await _reader.ReadAsync(ch, 0, 128);
             var text = new string(ch, 0, count);
             Assert.Equal("", text);
+        }
+
+        public static Socket CreateConnectedLoopbackSocket(int port)
+        {
+            var socket = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp);
+            if (PlatformApis.IsWindows)
+            {
+                const int SIO_LOOPBACK_FAST_PATH = -1744830448;
+                var optionInValue = BitConverter.GetBytes(1);
+                try
+                {
+                    socket.IOControl(SIO_LOOPBACK_FAST_PATH, optionInValue, null);
+                }
+                catch
+                {
+                    // If the operating system version on this machine did
+                    // not support SIO_LOOPBACK_FAST_PATH (i.e. version
+                    // prior to Windows 8 / Windows Server 2012), handle the exception
+                }
+            }
+            socket.Connect(new IPEndPoint(IPAddress.Loopback, port));
+            return socket;
         }
     }
 }
